Consolidate wildcard matching function to msgbus
